OBPIH-6643 Add manual save on confirm page of final invoice for prepayment (#4814)
