+++ src/js/components/invoice/create/InvoiceOptionsForm.jsx
@@ -70,6 +70,7 @@ const InvoiceOptionsForm = ({
   disableSaveButton,
   updateInvoiceItem,
   canUpdateInvoiceItems,
+  save,
 }) => {
   const [isDropdownVisible, setIsDropdownVisible] = useState(false);
 
@@ -104,6 +105,22 @@ const InvoiceOptionsForm = ({
             />
           </span>
         </button>
+        {canUpdateInvoiceItems && (
+        <button
+          type="button"
+          className="btn btn-outline-secondary float-right btn-form btn-xs"
+          disabled={disableSaveButton}
+          onClick={() => save()}
+        >
+          <span>
+            <i className="fa fa-save pr-2" />
+            <Translate
+              id="react.default.button.save.label"
+              defaultMessage="Save"
+            />
+          </span>
+        </button>
+        )}
         <span className="mr-3">
           <div className="dropdown">
             <button
@@ -160,6 +177,7 @@ InvoiceOptionsForm.propTypes = {
   disableSaveButton: PropTypes.bool.isRequired,
   updateInvoiceItem: PropTypes.func.isRequired,
   canUpdateInvoiceItems: PropTypes.bool.isRequired,
+  save: PropTypes.func.isRequired,
 };
 
 export default InvoiceOptionsForm;
